+++ bs-config.php
@@ -1,5 +1,7 @@
 <?php
 
+/* Site language */
 define( 'BS_LANG', 'fr_FR' );
 
+/* Site language */
 define( 'BS_NAME', 'BackStarter Demo' );