+++ views/test.php
@@ -1,6 +1,6 @@
 <?php
 echo $this->getModule( 'header' );
 ?>
-<h1>Hello World !</h1>
+<h1>Test page</h1>
 <?php
 echo $this->getModule( 'footer' );