+++ changelog.md
@@ -4,6 +4,14 @@ Changelog
 2013-11-23 : v 0.1
 ---
 
-- Use config file
-- Load page with infos
-- Display Hello World
+* Use config file
+* Load page with infos
+* Display Hello World
+
+2013-11-23 : v 0.2
+---
+
+* Multiple pages
+* HTTP Headers
+* 404 & Test templates
+* Default viewport